% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{connect}
\alias{connect}
\title{Entities}
\usage{
connect(data, from, to = NULL)
}
\arguments{
\item{data}{The data, as returned by \link[webhoser]{wh_collect}.}

\item{from, to}{Entities to build network.}
}
\value{
A \code{list} of length 2 containing \code{data.frame}s:
\itemize{
  \item{\code{nodes}: Name of \code{entity} and \code{n}umber of occurences.}
  \item{\code{edges}: The \code{source}, \code{target}, and \code{n}umber of edges.}
}
}
\description{
Create network from entities.
}
\details{
The returned nodes and edges form an \emph{undirected} graph.
}
\examples{
data("webhoser")

# co-mentions
graph <- webhoser \%>\%
  connect(entities.persons)

# unpack
c(nodes, edges) \%<-\% graph

# visualise
webhoser \%>\%
  connect(entities.persons, entities.locations) \%>\% 
  visualise()

}
