% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{net_con}
\alias{net_con}
\title{Connect}
\usage{
net_con(data, from, to = NULL, callback = NULL)
}
\arguments{
\item{data}{The data, as returned by \link[webhoser]{wh_collect}.}

\item{from, to}{Columns to build network.}

\item{callback}{Callback to apply to edges, a data.frame with columns 
\code{source}, \code{target}, \code{n} (number of occurences).}
}
\value{
A \code{list} of length 2 containing \code{data.frame}s:
\itemize{
  \item{\code{nodes}: Name of \code{entity} and \code{n}umber of occurences.}
  \item{\code{edges}: The \code{source}, \code{target}, and \code{n}umber of edges.}
}
}
\description{
Create network from webhoser data.
}
\details{
The returned nodes and edges form an \emph{undirected} graph.
}
\examples{
data("webhoser")

# co-mentions
graph <- webhoser \%>\%
  net_con(entities.persons)

# unpack
c(nodes, edges) \%<-\% graph

# visualise
webhoser \%>\%
  net_con(entities.persons, entities.locations) \%>\% 
  net_vis()
  
# callback
cb <- function(x){
  dplyr::filter(x, n > 3)
}

webhoser \%>\%
  net_con(
    entities.persons, 
    entities.locations, 
    callback = cb
  ) \%>\% 
  net_vis()

}
