% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{net_entities}
\alias{net_entities}
\title{Entities}
\usage{
net_entities(data, entity)
}
\arguments{
\item{data}{The data, as returned by \link[webhoser]{wh_collect}.}

\item{entity}{Entity to build network.}
}
\value{
A \code{list} of length 2 containing \code{data.frame}s:
\itemize{
  \item{\code{nodes}: Name of \code{entity} and \code{n}umber of occurences.}
  \item{\code{edges}: The \code{source}, \code{target}, and \code{n}umber of edges.}
}
}
\description{
Create network from entities
}
\note{
The returned nodes and edges form an \emph{undirected} graph.
}
\examples{
data("webhoser")

webhoser \%>\%
  net_entities(entities.persons)

}
