% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nethoser-package.R
\docType{package}
\name{nethoser-package}
\alias{nethoser}
\alias{nethoser-package}
\title{nethoser: Networks for 'webhoser'}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Create various networks to understand results of 'webhoser'.
}
\author{
\strong{Maintainer}: John Coene \email{jcoenep@gmail.com}

}
\keyword{internal}
