% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{net_vis}
\alias{net_vis}
\alias{net_viz}
\title{Visualise}
\usage{
net_vis(graph)

net_viz(graph)
}
\arguments{
\item{graph}{A list as returned by \code{\link{net_con}}.}
}
\description{
Visualise nethoser network.
}
\examples{
data("webhoser")

# make network
webhoser \%>\%
  net_con(thread.site, entities.persons) \%>\% 
  net_vis()

}
