% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{webhoser}
\alias{webhoser}
\title{R Programming Articles}
\format{An object of class \code{data.frame} with 272 rows and 42 columns.}
\source{
\url{https://webhose.io}
}
\usage{
webhoser
}
\description{
A sample of data returned by the \link[webhoser]{wh_collect} function containing
news articles on the R programming language.
}
\keyword{datasets}
