% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize}
\alias{visualize}
\alias{visualise}
\title{Visualise}
\usage{
visualize(graph)

visualise(graph)
}
\arguments{
\item{graph}{A list as returned by \code{\link{connect}}.}
}
\description{
Visualise network.
}
\examples{
data("webhoser")

# make network
webhoser \%>\%
  connect(thread.site, entities.persons) \%>\% 
  visualize()

}
